Extract triggered email functions into their own subclass
